Rename service file to todoService
